refactor: follow 도메인 분리
