+++ src/main/java/com/example/sns/follow/domain/Follow.java
@@ -1,5 +1,6 @@
-package com.example.sns.member.domain;
+package com.example.sns.follow.domain;
 
+import com.example.sns.member.domain.Member;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
