+++ src/main/java/com/example/sns/member/domain/Member.java
@@ -2,6 +2,7 @@ package com.example.sns.member.domain;
 
 import com.example.sns.auth.application.dto.OAuthUserInfoDto;
 import com.example.sns.common.entity.BaseTimeEntity;
+import com.example.sns.follow.domain.Follow;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
